+++ modules/app-desktop/src/main/kotlin/org/archivekeep/app/desktop/ui/dialogs/repository/operations/indexupdate/UpdateIndexOperationDialog.kt
@@ -31,6 +31,7 @@ import org.archivekeep.app.desktop.ui.components.FileManySelect
 import org.archivekeep.app.desktop.ui.components.ItemManySelect
 import org.archivekeep.app.desktop.ui.components.dialogs.operations.DialogOperationControlButtons
 import org.archivekeep.app.desktop.ui.components.dialogs.operations.DialogOperationControlState
+import org.archivekeep.app.desktop.ui.components.dialogs.operations.ExecutionErrorIfPresent
 import org.archivekeep.app.desktop.ui.components.dialogs.operations.toDialogOperationControlState
 import org.archivekeep.app.desktop.ui.components.operations.IndexUpdatePreparationProgress
 import org.archivekeep.app.desktop.ui.components.operations.LocalIndexUpdateProgress
@@ -189,6 +190,7 @@ class UpdateIndexOperationDialog(
                 )
                 Spacer(Modifier.height(4.dp))
                 ScrollableLogTextInDialog(operationState.log)
+                ExecutionErrorIfPresent(operationState.state)
             }
         }
     }